requirements.txt
flask
flask-sqlalchemy
google-generativeai
pypdf2
requests
python-dotenv